+++ CHANGELOG.md
@@ -12,6 +12,11 @@ pubspec.yaml version does not match
 ### Changed
 
 - Fix issues with CHANGELOG.md
+## 12.7.1 - 2026-08-04
+
+### Removed
+
+- Remove duplicate upgrade calls from publish workflow
 
 ## 12.7.0 - 2026-08-04